django_app/requirements.txt
aioredis==1.3.1
altgraph==0.17.2
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
billiard==3.6.4.0
bottle==0.12.21
bottle-websocket==0.2.9
celery==5.2.3
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-redis==3.3.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
daphne==3.0.2
Django==3.2.10
django-appconf==1.0.5
django-celery-beat==2.2.1
django-celery-email==3.0.0
django-cors-headers==3.10.1
django-polymorphic==3.1.0
django-rest-swagger==2.2.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.0.0
drf-jwt==1.19.1
drf-spectacular==0.21.1

Eel==0.14.0
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
importlib-resources==5.9.0
incremental==21.3.0
inflection==0.5.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.7.2
kombu==5.2.4
macholib==1.16
MarkupSafe==2.1.1
msgpack==1.0.4
openapi-codec==1.3.2
prompt-toolkit==3.0.30
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyinstaller==5.2
pyinstaller-hooks-contrib==2022.8
PyJWT==1.7.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.1
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.10
vine==5.0.0
wcwidth==0.2.5
whichcraft==0.6.1
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
psycopg2-binary==2.9.2
mysqlclient>=2.0.1,<2.1.0
Redis
django_smtp_ssl
django-mjml==0.11.0
exponent_server_sdk==2.0.0
rollbar==0.16.3
pymysql>=0.10.1,<0.11.0